requirements.txt
uvicorn
fastapi
jinja2
google-generativeai
python-multipart
starlette